requirements.txt
fastapi
uvicorn
pytesseract
Pillow
opencv-python
numpy
python-multipart
